refactor(Global): Refactored a bunch of code naming conventions and locations
