+++ components/tailored/core/sidebar/controls/Controls.vue
@@ -4,7 +4,7 @@
 import Vue from 'vue'
 
 import { mapState } from 'vuex'
-import { Sounds } from '~/utilities/SoundManager/SoundManager'
+import { Sounds } from '~/libraries/SoundManager/SoundManager'
 
 export default Vue.extend({
   computed: {
