+++ components/tailored/marketplace/Marketplace.html
@@ -22,7 +22,7 @@
                     type="primary"
                     :icon="{ style: 'far', name: 'grin-tongue-wink' }"
                     outlined
-                    :action="testAction"
+                    :action="() => {}"
                 />
                 <InteractablesButton
                     text="Themes"
@@ -30,7 +30,7 @@
                     type="primary"
                     :icon="{ style: 'far', name: 'palette' }"
                     outlined
-                    :action="testAction"
+                    :action="() => {}"
                 />
                 <InteractablesButton
                     text="NFTs"
@@ -38,7 +38,7 @@
                     type="primary"
                     :icon="{ style: 'far', name: 'cubes' }"
                     outlined
-                    :action="testAction"
+                    :action="() => {}"
                 />
             </TailoredMarketplaceButtongroup>
         </div>
