+++ utilities/README.md
@@ -1,0 +1,7 @@
+# Utilities
+
+Utilities are methods or functions who complete a simple task and do not drastically interface with external APIs
+
+## Other Considerations
+
+You should consider upgrading a utility to a library if the utility becomes complex and has many moving parts.
